LayoutTests/fast/forms/ios/inputmode-change-update-keyboard-expected.txt

This tests updating inputmode of an input element from "none" to "text". The software keyboard should be updated.
To manually test, focus the input element below. The software keyboard should show up after 3 seconds

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


inputmode="text"
PASS textHeight = keyboardRect.height; keyboardRect.height > 0 is true

inputmode="number"
PASS keyboardRect.height is not textHeight
PASS successfullyParsed is true

TEST COMPLETE
